requirements.txt
cryptography==1.2.3
Flask==0.12
Flask-Cors==3.0.2
Flask-OAuthlib==0.9.3
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
html5lib==0.999
httplib2==0.9.1
idna==2.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
PyJWT==1.4.2
SQLAlchemy==1.1.5
Werkzeug==0.11.15
passlib==1.7.1
cymysql==0.8.9
Flask-Session==0.3.0